Module Task.File

val read_text : File.t -> (string -> 'm) -> ('m, empty) t

read_text file f

Read the contents of file into a string result and produce the message f result when the reading is finished.