Module Task.Http

type error = [

| `Status of int
(*

0: no internet, server not found, timeout, ...

401: bad request

403: forbidden

404: page not found

...

*)

| `No_json
(*
Resource is not a valid json file
*)

| `Decode
(*
Resource is a valid json file, but the decoder could not decode the corresponding javascript object.
*)

]

module Body : sig ... end

module Expect : sig ... end

val request : string -> string -> (string * string) list -> Body.t -> 'a Expect.t -> ('a, error) t

request method url headers body expect

Make an http method request to url with headers and body. expect specifies the expected response format.

This is the most general http request function. See also the more specific functions text and json.

Example:

    let user = Value.(record [| ("username", string "Bob") |]) in
    request "PUT" "/users" [] (Body.json user) (Expect.string)
    |> Command.attempt (fun result ->
        match result with
        | Ok _ ->
            GotUserCreated
        | Error _ ->
            GotError "failed to create user")

val text : string -> string -> (string * string) list -> string -> (string, error) t

text method url headers body

Make an http method request to url with headers and a string as the body. Expect a string as the response.

Method is one of GET, POST, DELETE, ... .

The headers and the body can be empty. The Content-Type header is automatically set to text/plain.

Example:

    text "PUT" "/users" [] "Bob"
    |> Command.attempt (fun result ->
        match result with
        | Ok _ ->
            GotUserCreated
        | Error _ ->
            GotError "failed to create user")

val json : string -> string -> (string * string) list -> Value.t option -> 'a Decoder.t -> ('a, error) t

json method url headers body decoder

Make an http method request to url with headers and an optional json value as the body. Expect a json value as the response which will be decoded by decoder.

The headers can be empty. The Content-Type header is automatically set to application/json if body is not None.

Example:

    let decoder = Decoder.array Decoder.string in
    json "GET" "/users" [] None decoder
    |> Command.attempt (fun result ->
        match result with
        | Ok usernames -> (* the usernames were successfully decoded
                             into a string array *)
            GotUsers (Array.to_list usernames)
        | Error _ ->
            GotError "failed to obtain users")