Module Time.Zone

Time zone

A time zone is the offset in minutes from utc. Time zones westward of utc get a positive offset, eastward of utc a negative offset.

type t

val utc : t

val make : int -> t

make offset Time zone offset minutes westward of utc.

make (-60) is the zone of central european winter time. It is one hour eastward of utc.

val offset : t -> int

offset zone The offset of zone in minutes westward of utc.