Module Builder.Query

Construct queries

type t

The query builder type

val raw : string -> string -> t

raw key s creates a query parameter with the given key and string value s without percent-encoding either.

This is useful for including readable unicode graphemes in the URL query. See absolute for examples.

val string : string -> string -> t

string key s creates a query parameter with the given key and string value s..

Special characters in the key and in the value, like ?, / or non-ASCII characters will be automatically escaped using percent-encoding. See absolute for examples.

val int : string -> int -> t

int key i creates a URL query parameter with the given key and integer value i.

See absolute for examples.